+++ pkg/deployment/agency/state/shards.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,23 +18,6 @@
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
 
-package agency
+package state
 
-type JobPhase string
-
-var (
-	JobPhaseUnknown  JobPhase = ""
-	JobPhaseToDo     JobPhase = "ToDo"
-	JobPhasePending  JobPhase = "Pending"
-	JobPhaseFailed   JobPhase = "Failed"
-	JobPhaseFinished JobPhase = "Finished"
-)
-
-type JobID string
-
-type Jobs map[JobID]Job
-
-type Job struct {
-	Type   string `json:"type,omitempty"`
-	Reason string `json:"reason,omitempty"`
-}
+type Shards map[string]Servers