+++ pkg/deployment/reconcile/action_wait_for_member_in_sync.go
@@ -24,7 +24,7 @@ import (
 	"context"
 
 	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
-	"github.com/arangodb/kube-arangodb/pkg/deployment/agency"
+	"github.com/arangodb/kube-arangodb/pkg/deployment/agency/state"
 	"github.com/arangodb/kube-arangodb/pkg/util"
 )
 
@@ -95,7 +95,7 @@ func (a *actionWaitForMemberInSync) checkCluster() (bool, error) {
 			return false, nil
 		}
 
-		notInSyncShards := agency.GetDBServerShardsNotInSync(agencyState, agency.Server(a.MemberID()))
+		notInSyncShards := state.GetDBServerShardsNotInSync(agencyState, state.Server(a.MemberID()))
 
 		if len(notInSyncShards) > 0 {
 			a.log.Str("mode", "cluster").Str("member", a.MemberID()).Int("shard", len(notInSyncShards)).Info("DBServer contains not in sync shards")