+++ pkg/deployment/reconcile/plan_builder_scale.go
@@ -26,7 +26,7 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/apis/deployment"
 	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 	"github.com/arangodb/kube-arangodb/pkg/deployment/actions"
-	"github.com/arangodb/kube-arangodb/pkg/deployment/agency"
+	"github.com/arangodb/kube-arangodb/pkg/deployment/agency/state"
 	"github.com/arangodb/kube-arangodb/pkg/deployment/reconcile/shared"
 	"github.com/arangodb/kube-arangodb/pkg/deployment/reconciler"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
@@ -178,7 +178,7 @@ func getCleanedServer(ctx reconciler.ArangoAgencyGet) api.MemberToRemoveSelector
 	return func(m api.MemberStatusList) (string, error) {
 		if a, ok := ctx.GetAgencyCache(); ok {
 			for _, member := range m {
-				if a.Target.CleanedServers.Contains(agency.Server(member.ID)) {
+				if a.Target.CleanedServers.Contains(state.Server(member.ID)) {
 					return member.ID, nil
 				}
 			}
@@ -191,7 +191,7 @@ func getToBeCleanedServer(ctx reconciler.ArangoAgencyGet) api.MemberToRemoveSele
 	return func(m api.MemberStatusList) (string, error) {
 		if a, ok := ctx.GetAgencyCache(); ok {
 			for _, member := range m {
-				if a.Target.ToBeCleanedServers.Contains(agency.Server(member.ID)) {
+				if a.Target.ToBeCleanedServers.Contains(state.Server(member.ID)) {
 					return member.ID, nil
 				}
 			}
@@ -213,13 +213,13 @@ func getDbServerWithLowestShards(ctx reconciler.ArangoAgencyGet, g api.ServerGro
 
 		dbServersShards := a.ShardsByDBServers()
 		for _, member := range m {
-			if _, ok := dbServersShards[agency.Server(member.ID)]; !ok {
+			if _, ok := dbServersShards[state.Server(member.ID)]; !ok {
 				// member is not in agency cache, so it has no shards
 				return member.ID, nil
 			}
 		}
 
-		var resultServer agency.Server = ""
+		var resultServer state.Server = ""
 		var resultShards int
 
 		for server, shards := range dbServersShards {